+++ src/helpers.ts
@@ -1,1 +1,6 @@
-export type StripNever<T> = T extends never ? never : T
+import { Err, Ok, Result } from "./Result"
+
+type UnwrapOk<T> = T extends Ok<infer U> ? U : never
+type UnwrapErr<T> = T extends Err<infer E> ? E : never
+
+export type Merged<T> = Result<UnwrapOk<T>, UnwrapErr<T>>